# openapi/messaging_openapi.json
{"openapi":"3.1.0","info":{"title":"Messaging Microservice","description":"Handles messaging and threads for the Neighborhood Exchange app","version":"1.0.0"},"paths":{"/threads/":{"post":{"tags":["Threads"],"summary":"Create Thread","operationId":"create_thread_threads__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ThreadCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ThreadRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/threads/{thread_id}":{"get":{"tags":["Threads"],"summary":"Get Thread","operationId":"get_thread_threads__thread_id__get","parameters":[{"name":"thread_id","in":"path","required":true,"schema":{"type":"string","title":"Thread Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ThreadRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/threads/{thread_id}/messages":{"post":{"tags":["Messaging"],"summary":"Send Message","operationId":"send_message_threads__thread_id__messages_post","parameters":[{"name":"thread_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Thread Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["Messaging"],"summary":"Get Messages","operationId":"get_messages_threads__thread_id__messages_get","parameters":[{"name":"thread_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Thread Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MessageRead"},"title":"Response Get Messages Threads  Thread Id  Messages Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MessageCreate":{"properties":{"sender_id":{"type":"string","format":"uuid","title":"Sender Id"},"content":{"type":"string","maxLength":1000,"minLength":1,"title":"Content"}},"type":"object","required":["sender_id","content"],"title":"MessageCreate","example":{"content":"Hey! Are you still selling the mirror?","sender_id":"123e4567-e89b-12d3-a456-426614174000"}},"MessageRead":{"properties":{"sender_id":{"type":"string","format":"uuid","title":"Sender Id","description":"UUID of the sender"},"content":{"type":"string","minLength":1,"title":"Content","description":"Content of the message."},"time_stamp":{"type":"string","format":"date-time","title":"Time Stamp","description":"Time when the message was sent (UTC)."},"id":{"type":"string","format":"uuid","title":"Id"},"thread_id":{"type":"string","format":"uuid","title":"Thread Id","description":"Thread this message belongs to"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["sender_id","content","thread_id"],"title":"MessageRead","example":{"content":"Got it, thank you!","created_at":"2025-11-10T12:34:56Z","id":"550e8400-e29b-41d4-a716-446655440000","sender_id":"123e4567-e89b-12d3-a456-426614174000","thread_id":"123e4567-e89b-12d3-a456-426614173000","time_stamp":"2025-11-10T12:34:56Z","updated_at":"2025-11-10T12:45:00Z"}},"ThreadCreate":{"properties":{"author_id":{"type":"string","format":"uuid","title":"Author Id","description":"User starting the conversation."},"participant_id":{"type":"string","format":"uuid","title":"Participant Id","description":"Other user in the DM thread."}},"type":"object","required":["author_id","participant_id"],"title":"ThreadCreate","example":{"author_id":"a1b2c3d4-e5f6-7890-1234-567890abcdef","participant_id":"bb12cc34-dd56-ee78-9900-aabbccddeeff"}},"ThreadRead":{"properties":{"author_id":{"type":"string","format":"uuid","title":"Author Id","description":"The user who created the thread."},"participant_id":{"type":"string","format":"uuid","title":"Participant Id","description":"The other person in the conversation."},"thread_id":{"type":"string","format":"uuid","title":"Thread Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["author_id","participant_id","thread_id"],"title":"ThreadRead","description":"Returned to the client when fetching thread info.","example":{"author_id":"a1b2c3d4-e5f6-7890-1234-567890abcdef","created_at":"2025-11-10T12:34:56Z","participant_id":"bb12cc34-dd56-ee78-9900-aabbccddeeff","thread_id":"550e8400-e29b-41d4-a716-446655440000","updated_at":"2025-11-10T12:35:56Z"}},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}